feat: there are too many Sablier events, get them from logs instead, update roles stream keys to remove duplicate info

+++ packages/decent-ponder/ponder.schema.ts
@@ -111,7 +111,7 @@ export const stream = onchainTable(
     amount: bigint(),
     start: integer(),
     cliff: integer(),
-    end: bigint(),
+    end: integer(),
     cancelable: boolean(),
     transferable: boolean(),
   },
